+++ internal/pkg/mlrval/mlrmap_accessors_test.go
@@ -21,7 +21,7 @@ func TestPutReference(t *testing.T) {
 
 	assert.True(t, mlrmap.Has("a"))
 	assert.False(t, mlrmap.Has("b"))
-	assert.Equal(t, 1, mlrmap.FieldCount)
+	assert.Equal(t, int64(1), mlrmap.FieldCount)
 
 	read := mlrmap.Get("b")
 	assert.Nil(t, read)
@@ -30,7 +30,7 @@ func TestPutReference(t *testing.T) {
 	assert.NotNil(t, read)
 	intval, ok := read.GetIntValue()
 	assert.True(t, ok)
-	assert.Equal(t, 1, intval)
+	assert.Equal(t, int64(1), intval)
 
 	key2 := "b"
 	val2 := FromBool(true)
@@ -38,7 +38,7 @@ func TestPutReference(t *testing.T) {
 
 	assert.True(t, mlrmap.Has("a"))
 	assert.True(t, mlrmap.Has("b"))
-	assert.Equal(t, 2, mlrmap.FieldCount)
+	assert.Equal(t, int64(2), mlrmap.FieldCount)
 
 	read = mlrmap.Get("a")
 	assert.NotNil(t, read)
